pyproject.toml
[project]
name = "ai-test-automation-proto"
version = "0.1.0"
description = "Prototype: PDF RAG (Chroma+Ollama) + controlled web scraping (Playwright) + simple pipeline CLI."
requires-python = ">=3.11"
dependencies = [
  # RAG core
  "chromadb>=0.5.0",
  "langchain>=0.2.0",
  "langchain-community>=0.2.0",
  "langchain-text-splitters>=0.2.0",
  "langchain-chroma>=0.1.0",
  "langchain-ollama>=0.1.0",
  "pypdf>=4.0.0",

  # CLI + logging
  "typer>=0.12.0",
  "rich>=13.0.0",

  # Web
  "playwright>=1.45.0",
  "lxml>=5.0.0",
  "cssselect>=1.2.0",
  "googlesearch-python>=1.2.3",
]

[project.scripts]
atp = "atp.cli:app"

[tool.uv]
package = true
